requirements.txt
torch  # Recommended: 2.8.0
faiss-gpu
transformers>=4.36.0
peft>=0.7.0
lightning
sentencepiece>=0.1.99
gradio>=4.9.0
matplotlib>=3.8.0
nltk>=3.8.1
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
regex>=2023.10.0
scikit-learn>=1.3.0
scipy>=1.10.0
tensorboard>=2.15.0
tqdm>=4.66.0
